+++ src/test/java/com/networknt/client/orchestration/ParallelRequestsTest.java
@@ -4,10 +4,7 @@ package com.networknt.client.orchestration;
 import com.networknt.client.model.HttpVerb;
 import com.networknt.petstore.handler.TestServer;
 import com.networknt.petstore.model.Pet;
-import org.junit.Assert;
-import org.junit.Before;
-import org.junit.ClassRule;
-import org.junit.Test;
+import org.junit.*;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;